Extract interop envelope interfaces
Add a new Contracts/IInteropEnvelope.ts file containing InteropEnvelopeV1, ParsedInteropMessage, and ParseError interfaces, and update Interop/InteropEnvelopeProtocol.ts to import these types instead of declaring them inline. Consolidates type definitions to a single contract file to reduce duplication and improve reuse; no functional changes to protocol logic.
